requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.4.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
cssselect2==0.8.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.1
django-cloudinary-storage==0.3.0
dotenv==0.9.9
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.12.0
gitignore_parser==0.1.13
gunicorn==23.0.0
h11==0.16.0
idna==3.11
ImageHash==4.3.2
Jinja2==3.1.6
jiter==0.11.1
Levenshtein==0.27.3
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.3.4
orjson==3.11.4
packaging==25.0
pillow==11.3.0
plotly==5.24.1
propcache==0.4.1
psycopg2-binary==2.9.11
pyarrow==18.1.0
pycparser==2.23
pydantic_core==2.41.4
pydyf==0.11.0
PyJWT==2.11.0
pylibdmtx==0.1.10
pyphen==0.17.2
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-Levenshtein==0.27.3
python-multipart==0.0.20
pytz==2024.2
PyWavelets==1.9.0
RapidFuzz==3.14.3
regex==2026.2.28
reportlab==4.4.4
requests==2.32.5
scipy==1.17.1
segno==1.6.6
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.3
treepoem==3.28.0
twilio==9.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchdog==3.0.0
wcwidth==0.6.0
weasyprint==66.0
webencodings==0.5.1
websockets==12.0
whitenoise==6.11.0
yarl==1.22.0
zopfli==0.2.3.post1
